+++ api/src/models/community_business.ts
@@ -399,6 +399,7 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
       .returning([
         'visit_activity_id AS id',
         'visit_activity_name AS name',
+        'visit_activity_category_id AS category_id',
         'monday',
         'tuesday',
         'wednesday',
@@ -409,7 +410,12 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
         'created_at AS createdAt',
         'modified_at as modifiedAt',
       ]);
-    return res;
+
+    const [{ category }]: { category: string }[] = await client('visit_activity_category')
+      .select('visit_activity_category_name AS category')
+      .where({ visit_activity_category_id: res.category_id })
+
+    return { ...omit(['category_id'], res) as Omit<VisitActivity, 'category'>, category };
   },
 
   async updateVisitActivity (client, visitActivity) {
@@ -430,6 +436,7 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
       .returning([
         'visit_activity_id AS id',
         'visit_activity_name AS name',
+        'visit_activity_category_id AS cat_id',
         'monday',
         'tuesday',
         'wednesday',
@@ -441,7 +448,17 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
         'modified_at AS modifiedAt',
       ]);
 
-    return res ? { ...res, category: visitActivity.category } : null;
+    if (!res) {
+      return null;
+    }
+
+    const [{ category }]: { category: string }[] = await client('visit_activity_category')
+      .select('visit_activity_category_name AS category')
+      .where({ visit_activity_category_id: res.cat_id });
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { cat_id, ...activity } = res;
+    return { ...activity, category };
   },
 
   async deleteVisitActivity (client, id) {
@@ -586,10 +603,10 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
       , queryMatchOnColumnNames)
         .then((rows) => {
           const gender: Dictionary<number> = rows
-            .reduce((acc: Dictionary<number>, row: {gender: string; count: number}) => {
+            .reduce((acc: Dictionary<number>, row: { gender: string; count: number }) => {
               acc[row.gender] = Number(row.count);
               return acc;
-            } , {});
+            }, {});
           return { gender };
         }),
 
@@ -607,10 +624,10 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
         .groupBy('visit_activity.visit_activity_name'), queryMatchOnColumnNames)
         .then((rows) => {
           const visitActivity = rows
-            .reduce((acc: Dictionary<number>, row: {activity: string; count: number}) => {
+            .reduce((acc: Dictionary<number>, row: { activity: string; count: number }) => {
               acc[row.activity] = Number(row.count);
               return acc;
-            } , {});
+            }, {});
           return { visitActivity };
         }),
 
@@ -627,7 +644,7 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
             'FROM visit_log ' +
             'INNER JOIN user_account ON user_account.user_account_id = visit_log.user_account_id')
       )
-      // TODO: generate case statements based on supplied query
+        // TODO: generate case statements based on supplied query
         .innerJoin(
           'visit_activity',
           'visit_activity.visit_activity_id',
@@ -639,10 +656,10 @@ export const CommunityBusinesses: CommunityBusinessCollection = {
         .from('age_group_table'), ageQuery)
         .then((rows) => {
           const age = rows
-            .reduce((acc: Dictionary<number>, row: {ageGroup: string; count: number}) => {
+            .reduce((acc: Dictionary<number>, row: { ageGroup: string; count: number }) => {
               acc[row.ageGroup] = Number(row.count);
               return acc;
-            } , {});
+            }, {});
           return { age };
         }),
       lastWeek: applyQueryModifiers(client('visit_log')
